Returns resources with user_visible not set to False
With this patch a resource is hidden for a user
only if the user_visible attribute is set to False.

This patch comes with a schema change which removes
the default value of the user_visible attribute.
The default value is going to be removed in order
to be able to distinguish if the user_visible attribute
has to be updated during update operation. With a
default value set to True we don't know whether the attribute
has changed or if it is just not set.

Change-Id: Ice912b463995c47684e171e1aea00424140d9004