config: do not read backref/children fields from cassandra if not needed
from the bug descr:
in cassandra_<type>_read() implementation of autogen code, all the columns
of an object are read from DB and filtered by field_names by iteration.

In scale situations this is non-performant as project, ipam, sg etc. have
many backrefs and most common case is to not need backref/children of these objects.

So fix autogen code to check required fields against the set of backref + children
and read-in with column_start if not only props/ref are needed.

Change-Id: Ie96744e829f67409c6cb76deecb4e7117ac61d5d
Closes-Bug: #1421752