config-perf: restrict db fetch to only obj_uuids when specified
In ifmap_frontend.py autogenerator, for the cassandra_<type>_list method,
any passed obj_uuids is used as a filter along with parent_id or back_ref_id.
obj_uuids passed on its own is ignored. Fix this and if only obj_uuids
is passed restrict the fetch from database to only the specified object uuids.

Change-Id: Ibe688d3832b6d667c4fd43777360d46e5d60a501
Closes-Bug: #1431035