Fix get methods for children and back refs in resource_client
While getting children and back refs, we were returning error if svr_conn was
None. It is not set in the newly introduced vnc_db api. So, don't return
None if the field is already present in the object.

Also, in cleanUp method for fixtures, do not find back-refs and children.
Instead, just try to delete and catch RefsExistError

Change-Id: I7cbd60f464223ca982d216ad14d6102bb549f1a1
Partial-Bug: 1470621