Fix the logic regarding reference updates.
References can be updated via the PUT operation on the object
or via the update-ref URI. Ensure that they are only done via
one of these mechanisms (not both on the same update).

Reference clear operations do not perform a read. Do not store
the current value of the reference list when it is unknown.

Fix the copying of the data used to compare references.

Change-Id: Ie24588d315f0c70f48040258c1b270b0b84a85d5
Closes-bug: #1541654