Distinguish resource type with dash to underscore
That fix introduce a distinction between resource type name with dash
to resource type name with underscore in order to not have to replace
dash by underscore and conversely. Dash version is principally use in
the rest API and in the ifamp graph, underscore one is used in code and
in database.

Partial-bug: #1586327
Change-Id: I535aabe08a2230e3b7c0e08b9d59472fdd1e3b7c