Distinguish resource type with dash to underscore
That fix introduce a distinction between resource type name with dash
to resource type name with underscore in order to not have to replace
dash by underscore and conversely. Dash version is principally use in
the rest API and in the ifmap graph, underscore one is used in code and
in database.
That patch adds a new list of tuples that contains all resource types in
both way (underscore and dash) named 'all_resource_type_tuples'. It
does not change the already existing list 'all_resource_types' to not
break compatibility until I8802a668b60cf173c11787ed86233b393a59ae1e was
merged.

Partial-bug: #1586327
Change-Id: I535aabe08a2230e3b7c0e08b9d59472fdd1e3b7c