Do not generate default arguments in client library
Currently, python version of vnc api client generates default
values of all paraemeters. This causes a problem when RBAC is
enabled for individual fields. With this change, we will not
populate default values on the client side. As a subsequent commit,
we will populate the default values on the server side.

Change-Id: Iadbbfb7be20b44102cd52e1fea42a23454df7ae5
Partial-Bug: 1699558