In JSON Parse, check for parsed entity type before accessing members
In generated code, when an object is accessed, ensure that the parsed
entity is indeed an object. Otherwise rapidjson code can potentially
assert and fail, crashing the process.

In one of the DB dumps, prop:ecmp_hashing_include_fields was encoded as
an empty array "[]" for a couple of UUIDs, when it should have always
been encoded as an object ("{}") instead.

Tesed by successfully loading the "bad" DB with config_json_parser_test.

Change-Id: I2dce03d2523bc1319fd5f1c508d2d98cbce44c22