Update properties only if it is marked 'U'
Right now we are updating resource properties without checking
whether it should be updated or not and hence we see this issue.

To fix this, I have added a check to see if the property is marked as
'U' in property operation and then only generate code for those
properties in handle_update.

Change-Id: If1b0630d5e3c00ef2429b99cfb8456da9f82a16f
closes-jira-bug: CEM-4958